Scenes are changeable thx to phaser now
