+++ app/javascript/channels/game.js
@@ -1,9 +1,19 @@
 import Phaser from "phaser"
+import { Waiting } from "../scenes/waiting"
+import { Login } from "../scenes/login"
 
-const config = {
+var config = {
+  type: Phaser.AUTO,
   width: 800,
-  height: 500,
-  type: Phaser.AUTO
-}
+  height: 600,
+  parent: "phaser-example",
+  physics: {
+      default: 'arcade',
+      arcade: {
+          gravity: { y: 200 }
+      }
+  },
+  scene: [Waiting, Login]
+};
 
-const game = new Phaser.game(config);
+const game = new Phaser.Game(config);
