+++ app/views/pages/home.html.erb
@@ -22,106 +22,3 @@
   <%= render "pages/options" %>
 </div>
 
-<script>
-var config = {
-  type: Phaser.AUTO,
-  width: 800,
-  height: 600,
-  physics: {
-    default: 'arcade',
-    arcade: {
-        gravity: { y: 300 },
-        debug: false
-    }
-  },
-  scene: {
-      preload: preload,
-      create: create,
-      update: update
-  }
-};
-
-var game = new Phaser.Game(config);
-var platforms;
-var player;
-
-function preload ()
-{
-  this.load.image('sky', '<%= image_path('sky.png') %>');
-  this.load.image('ground', '<%= image_path('platform.png')  %>');
-  this.load.image('star', '<%= image_path('star.png')  %>');
-  this.load.image('bomb', '<%= image_path('bomb.png')  %>');
-  this.load.spritesheet('dude', 
-      '<%= image_path('dude.png')  %>',
-      { frameWidth: 32, frameHeight: 48 }
-  );
-}
-
-function create ()
-{
-this.add.image(400, 300, 'sky');
-
-platforms = this.physics.add.staticGroup();
-
-platforms.create(400, 568, 'ground').setScale(2).refreshBody();
-
-platforms.create(600, 400, 'ground');
-platforms.create(50, 250, 'ground');
-platforms.create(750, 220, 'ground');
-
-player = this.physics.add.sprite(100, 450, 'dude');
-
-player.setBounce(0.2);
-player.setCollideWorldBounds(true);
-
-this.anims.create({
-    key: 'left',
-    frames: this.anims.generateFrameNumbers('dude', { start: 0, end: 3 }),
-    frameRate: 10,
-    repeat: -1
-});
-
-this.anims.create({
-    key: 'turn',
-    frames: [ { key: 'dude', frame: 4 } ],
-    frameRate: 20
-});
-
-this.anims.create({
-    key: 'right',
-    frames: this.anims.generateFrameNumbers('dude', { start: 5, end: 8 }),
-    frameRate: 10,
-    repeat: -1
-});
-this.physics.add.collider(player, platforms);
-cursors = this.input.keyboard.createCursorKeys();
-}
-
-function update ()
-{
-  if (cursors.left.isDown)
-  {
-      player.setVelocityX(-160);
-  
-      player.anims.play('left', true);
-  }
-  else if (cursors.right.isDown)
-  {
-      player.setVelocityX(160);
-  
-      player.anims.play('right', true);
-  }
-  else
-  {
-      player.setVelocityX(0);
-  
-      player.anims.play('turn');
-  }
-  
-  if (cursors.up.isDown && player.body.touching.down)
-  {
-      player.setVelocityY(-330);
-  }
-}
-
-  </script>
